Adds layermasks for targeting and occlusion.
Adds layermasks for targeting and camera occlusion. Now occlusion no longer relies on objects that shouldn't get in the way to be on the ignore raycast layer. They can be added to one of two layers, "No Occlusion" and "No Occlusion or Targetting", for the camera to ignore them in it's occlusion checking. The player is also on it's own layer and is ignored during targeting, allowing for an object which is behind the player or partially behind to still be targeted.